backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.4.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
pymysql==1.1.1
cryptography==42.0.8
email-validator==2.1.1
